FastSpringBoneBufferCombiner の責務を分割。
- FastSpringBoneCombinedBuffer 統合バッファ
- FastSpringBoneBufferCombiner 統合前バッファの登録管理
